Feature: Add configurable option to delete save on death, implement options persistence, and enhance shop UI interactions
Took 24 minutes

+++ Teal Way RPG/GeneralData/GameOver.cs
@@ -24,7 +24,10 @@ namespace Teal_Way_RPG.GeneralData
                 PKC("Press any key to start a new life...");
             }
 
-            SaveLoad.DeleteSaveFile();
+            if (Options.DeleteSaveOnDeath)
+            {
+                SaveLoad.DeleteSaveFile();
+            }
             NewGame.ResetCurrentProgress();
             Battle.RequestExitToMainMenu();
         }
